# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Box API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /auth/change-password:
    post:
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangePasswordRequest'}
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentUser'}
  /members:
    get:
      operationId: listMembers
      responses:
        '200':
          description: Organization accounts and external identities
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Member'}
    post:
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateAccountRequest'}
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Member'}
  /members/{memberId}:
    parameters:
      - name: memberId
        in: path
        required: true
        schema: {type: string, format: uuid}
    patch:
      operationId: updateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateAccountRequest'}
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Member'}
  /members/{memberId}/password:
    parameters:
      - name: memberId
        in: path
        required: true
        schema: {type: string, format: uuid}
    put:
      operationId: resetAccountPassword
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResetPasswordRequest'}
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Member'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: agentbox_session
...
  schemas:
...
    CurrentUser:
      type: object
      required: [id, login, displayName, role, status, mustChangePassword]
      properties:
        id: {type: string, format: uuid}
        login: {type: string}
        displayName: {type: string}
        role: {enum: [admin, user]}
        status: {enum: [active, disabled]}
        mustChangePassword: {type: boolean}
    Member:
      allOf:
        - {$ref: '#/components/schemas/CurrentUser'}
        - type: object
          required: [organizationId, hasPassword, createdAt]
          properties:
            organizationId: {type: string, format: uuid}
            hasPassword: {type: boolean}
            createdAt: {type: string, format: date-time}
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: {type: string}
        newPassword: {type: string, minLength: 8, maxLength: 72}
    CreateAccountRequest:
      type: object
      required: [username, displayName, role, password]
      properties:
        username: {type: string, minLength: 3, maxLength: 64}
        displayName: {type: string, minLength: 1, maxLength: 128}
        role: {enum: [admin, user]}
        password: {type: string, minLength: 8, maxLength: 72}
    UpdateAccountRequest:
      type: object
      properties:
        displayName: {type: string, minLength: 1, maxLength: 128}
        role: {enum: [admin, user]}
        status: {enum: [active, disabled]}
    ResetPasswordRequest:
      type: object
      required: [password]
      properties:
        password: {type: string, minLength: 8, maxLength: 72}
...